kickstart-tests: add a variant hostname login needle
text_console_login matches on the hostname, which the kickstart
'hostname' test changes (that's the point of the test). So add
a variant needle for that case. If we ever make a 'native'
openQA test that changes the hostname we should make sure to
use 'testhostname' so this needle is shared.