Run advisory_boot on UEFI too
We have a few update tests now which involve testing boot. So
we really should probably run them on UEFI as well as BIOS. I'd
like to do the installer image test this way too, but it's not
straightforward, as 'START_AFTER_TEST' filters by machine, so
you can't have a test run on the 'uefi' machine start after a
test run on the 'x86_64' machine...I can think of a way to work
around this but it's hideous, going to think harder before doing
it.

Signed-off-by: Adam Williamson <awilliam@redhat.com>